RCarb - Dose Rate Modelling of Carbonate-Rich Samples

The package provides a dose rate modelling for carbonate-rich samples in the context of trapped charged dating (e.g., luminescence dating) applications.

Details

Funding

Between 2018-2019, the work of Sebastian Kreutzer as maintainer of the package was supported by LabEx LaScArBxSK (ANR - n. ANR-10-LABX-52).

From 01/2020-04/2022, Sebastian Kreutzer received funding from the European Union’s Horizon 2020 research and innovation programme under the Marie Skłodowska-Curie grant agreement No 844457 (project: CREDit).

since 03/2023: Sebastian Kreutzer as maintainer of the package receives funding from the DFG Heisenberg programme No 505822867.

References

Kreutzer, S., Mauz, B., Martin, L., Mercier, N., 2019. “RCarb”: Dose Rate Modelling of Carbonate-Rich Samples - an Implementation of Carb in R -. Ancient TL 37, 1–8. doi:10.26034/la.atl.2019.533

This package bases on a 'MATLAB' programme with name 'Carb', details can be found the following references:

Mauz, B., Hoffmann, D., 2014. What to do when carbonate replaced water: Carb, the model for estimating the dose rate of carbonate-rich samples. Ancient TL 32, 24-32. doi:10.26034/la.atl.2014.481

Nathan, R.P., Mauz, B., 2008. On the dose-rate estimate of carbonate-rich sediments for trapped charge dating. Radiation Measurements 43, 14-25. doi:10.1016/j.radmeas.2007.12.012

Further reading

Nathan, R.P., 2010. Numerical modelling of environmental dose rate and its application to trapped-charge dating. DPhil thesis, St Hugh's College, Oxford. https://ora.ox.ac.uk/objects/ora:6421

See also

Author

Maintainer: Sebastian Kreutzer sebastian.kreutzer@uni-heidelberg.de (ORCID) [translator, data contributor]

Authors:

Roger P. Nathan [copyright holder]

Barbara Mauz (ORCID) [copyright holder]