Write table input template

This function creates a template table that can be used as input for the function model_DoseRate

write_InputTemplate(file = NULL, ...)

Arguments

file
character (optional): output path, if NULL nothing is written, but a template data.frame is returned.

...
additional arguments that can be passed to function write.table if file != NULL. Supported arguments are: sep, dec, fileEncoding`

Function version

0.1.0

See also

Examples

write_InputTemplate()

#>   SAMP_NAME    K  K_X    T  T_X   U  U_X U238 U238_X U234_U238 U234_U238_X WCI
#> 1   EXAMPLE 0.08 0.01 1.64 0.08 1.9 0.08    0      0         0           0  20
#>   WCI_X WCF WCF_X CC CC_X DIAM DIAM_X COSMIC COSMIC_X INTERNAL INTERNAL_X ONSET
#> 1     7   7     7 62    1  180     10   0.18     0.01        0          0   100
#>   ONSET_X FINISH FINISH_X DE DE_X
#> 1      10     40       10 98    9

# NOT RUN {
##Example with file output
write_InputTemplate(file = "~/Desktop/Input.csv")

# }