Reference data

Correction factors for beta and gamma radiation used for internal calculations. These values are used instead of the correction factors given in Aitken (1985) for the carbonate model.

Format

Reference_Data: list

NAME	TYPE	DIM	DESCRIPTION

DATAek	matrix	4 x 4	correction factors for electrons for water and carbonate to sediment mass ratio for K

DATAet	matrix	4 x 4	correction factors for electrons for water and carbonate to sediment mass ratio for Th

DATAet230	matrix	4 x 4	correction factors for electrons for water and carbonate to sediment mass ratio for Th-230

DATAeu	matrix	4 x 4	correction factors for electrons for water and carbonate to sediment mass ratio for U

DATAeu234	matrix	4 x 4	correction factors for electrons for water and carbonate to sediment mass ratio for U-234

DATAeu238	matrix	4 x 4	correction factors for electrons for water and carbonate to sediment mass ratio for U-238

DATApk	matrix	4 x 4	correction factors for photons for water and carbonate to sediment mass ratio for K

DATApt	matrix	4 x 4	correction factors for photons for water and carbonate to sediment mass ratio for T

DATApt230	matrix	4 x 4	correction factors for photons for water and carbonate to sediment mass ratio for Th-230

DATApu	matrix	4 x 4	correction factors for photons for water and carbonate to sediment mass ratio for U

DATApu234	matrix	4 x 4	correction factors for photons for water and carbonate to sediment mass ratio for U-234

DATApu238	matrix	4 x 4	correction factors for photons for water and carbonate to sediment mass ratio for U-238

mejdahl	data.frame	36 x 4	beta-dose attenuation values for quartz grains according to Mejdahl (1979)

Details

The reference values are used internally to account for: (1) grain size depend beta-attenuation factors (Mejdahl, 1979) and (2) to correct nuclide dependent beta and gamma radiation for water/carbonate proportions. The latter values are given as matrix and precise values are interpolated during the modelling process.

Different values quoted for U-238 and U-234 accounts for different activity ratios. For further details on the origin of these data we refer to Nathan & Mauz (2008) and Nathan (2010).

Version

0.1.0

References

Mejdahl, V., 1979. Thermoluminescence dating: beta-dose attenuation in quartz grains. Archaeometry 21, 61-72.

Nathan, R.P., Mauz, B., 2008. On the dose-rate estimate of carbonate-rich sediments for trapped charge dating. Radiation Measurements 43, 14-25. doi:10.1016/j.radmeas.2007.12.012

Nathan, R.P., 2010. Numerical modelling of environmental dose rate and its application to trapped-charge dating. DPhil thesis, St Hugh’s College, Oxford.

Further reading

Aitken, M.J., 1985. Thermoluminescence dating. Academic Press.

Examples

data(Reference_Data, envir = environment())
str(Reference_Data)

#> List of 13
#>  $ DATAek   : num [1:4, 1:4] 1 1.195 1.197 1.189 0.995 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:4] "1e-06" "0.04" "0.2" "1"
#>   .. ..$ : chr [1:4] "1e-06" "0.04" "0.2" "1"
#>  $ DATAet   : num [1:4, 1:4] 1 1.195 1.201 1.192 0.993 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:4] "1e-06" "0.04" "0.2" "1"
#>   .. ..$ : chr [1:4] "1e-06" "0.04" "0.2" "1"
#>  $ DATAet230: num [1:4, 1:4] 1 1.236 1.25 1.244 0.997 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:4] "1e-06" "0.04" "0.2" "1"
#>   .. ..$ : chr [1:4] "1e-06" "0.04" "0.2" "1"
#>  $ DATAeu   : num [1:4, 1:4] 1 1.193 1.199 1.187 0.992 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:4] "1e-06" "0.04" "0.2" "1"
#>   .. ..$ : chr [1:4] "1e-06" "0.04" "0.2" "1"
#>  $ DATAeu234: num [1:4, 1:4] 1 1.248 1.256 1.249 0.997 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:4] "1e-06" "0.04" "0.2" "1"
#>   .. ..$ : chr [1:4] "1e-06" "0.04" "0.2" "1"
#>  $ DATAeu238: num [1:4, 1:4] 1 1.187 1.189 1.177 0.993 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:4] "1e-06" "0.04" "0.2" "1"
#>   .. ..$ : chr [1:4] "1e-06" "0.04" "0.2" "1"
#>  $ DATApk   : num [1:4, 1:4] 1 1.06 1.06 1.04 1.02 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:4] "1e-06" "0.04" "0.2" "1"
#>   .. ..$ : chr [1:4] "1e-06" "0.04" "0.2" "1"
#>  $ DATApt   : num [1:4, 1:4] 1 1.023 1.015 0.987 1.039 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:4] "1e-06" "0.04" "0.2" "1"
#>   .. ..$ : chr [1:4] "1e-06" "0.04" "0.2" "1"
#>  $ DATApt230: num [1:4, 1:4] 1 0.26 0.252 0.236 1.494 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:4] "1e-06" "0.04" "0.2" "1"
#>   .. ..$ : chr [1:4] "1e-06" "0.04" "0.2" "1"
#>  $ DATApu   : num [1:4, 1:4] 1 1.015 1.007 0.976 1.044 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:4] "1e-06" "0.04" "0.2" "1"
#>   .. ..$ : chr [1:4] "1e-06" "0.04" "0.2" "1"
#>  $ DATApu234: num [1:4, 1:4] 1 0.208 0.206 0.202 1.527 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:4] "1e-06" "0.04" "0.2" "1"
#>   .. ..$ : chr [1:4] "1e-06" "0.04" "0.2" "1"
#>  $ DATApu238: num [1:4, 1:4] 1 0.814 0.792 0.716 1.16 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:4] "1e-06" "0.04" "0.2" "1"
#>   .. ..$ : chr [1:4] "1e-06" "0.04" "0.2" "1"
#>  $ mejdahl  :'data.frame':	36 obs. of  4 variables:
#>   ..$ D : num [1:36] 0.005 0.01 0.015 0.02 0.03 0.04 0.05 0.06 0.07 0.08 ...
#>   ..$ K : num [1:36] 0.0018 0.0035 0.0053 0.0071 0.0106 0.0141 0.0177 0.0212 0.0248 0.0283 ...
#>   ..$ Th: num [1:36] 0.0225 0.0366 0.0484 0.0582 0.0743 ...
#>   ..$ U : num [1:36] 0.012 0.0214 0.0296 0.0366 0.0475 0.0564 0.0642 0.0713 0.0779 0.084 ...

Reference_Data$DATAek

#>        1e-06   0.04    0.2      1
#> 1e-06 1.0000 0.9946 0.9936 0.9911
#> 0.04  1.1948 1.0944 1.0266 0.9985
#> 0.2   1.1973 1.1631 1.0922 1.0230
#> 1     1.1886 1.1808 1.1546 1.0850